Bulk Upload fix.
Ported over from dekkerglen/CubeCobra#2149.
